requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Flask-Cors==3.1.3
mediapipe==0.10.4
opencv-python==4.8.0.76
numpy==1.25.2
Pillow==10.0.0
scipy==1.11.2
